wip two bucket
This loops infinitely.. it gets quickly to a state where there are no legal moves and keeps cycling.

This is exactly as I suspected, that there will need to be an explosion at each step of alternative steps -- the subgroup execution method or whatever they call that. Although I am probably missing some much simpler way, because this problem is rated a medium.